+++ src/Symfony/Component/HttpClient/HttpClient.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\HttpClient;
 
+use Amp\Http\Client\Connection\ConnectionLimitingPool;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
@@ -29,12 +30,35 @@ final class HttpClient
      */
     public static function create(array $defaultOptions = [], int $maxHostConnections = 6, int $maxPendingPushes = 50): HttpClientInterface
     {
+        if ($amp = class_exists(ConnectionLimitingPool::class)) {
+            if (!\extension_loaded('curl')) {
+                return new AmpHttpClient($defaultOptions, null, $maxHostConnections, $maxPendingPushes);
+            }
+
+            // Skip curl when HTTP/2 push is unsupported or buggy, see https://bugs.php.net/77535
+            if (\PHP_VERSION_ID < 70217 || (\PHP_VERSION_ID >= 70300 && \PHP_VERSION_ID < 70304) || !\defined('CURLMOPT_PUSHFUNCTION')) {
+                return new AmpHttpClient($defaultOptions, null, $maxHostConnections, $maxPendingPushes);
+            }
+
+            static $curlVersion = null;
+            $curlVersion = $curlVersion ?? curl_version();
+
+            // HTTP/2 push crashes before curl 7.61
+            if (0x073d00 > $curlVersion['version_number'] || !(CURL_VERSION_HTTP2 & $curlVersion['features'])) {
+                return new AmpHttpClient($defaultOptions, null, $maxHostConnections, $maxPendingPushes);
+            }
+        }
+
         if (\extension_loaded('curl')) {
             if ('\\' !== \DIRECTORY_SEPARATOR || ini_get('curl.cainfo') || ini_get('openssl.cafile') || ini_get('openssl.capath')) {
                 return new CurlHttpClient($defaultOptions, $maxHostConnections, $maxPendingPushes);
             }
 
             @trigger_error('Configure the "curl.cainfo", "openssl.cafile" or "openssl.capath" php.ini setting to enable the CurlHttpClient', E_USER_WARNING);
+        }
+
+        if ($amp) {
+            return new AmpHttpClient($defaultOptions, null, $maxHostConnections, $maxPendingPushes);
         }
 
         return new NativeHttpClient($defaultOptions, $maxHostConnections);