+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * added `NoPrivateNetworkHttpClient` decorator
  * added `AmpHttpClient`, a portable HTTP/2 implementation based on Amp
  * added `LoggerAwareInterface` to `ScopingHttpClient` and `TraceableHttpClient`
+ * made `HttpClient::create()` return an `AmpHttpClient` when `amphp/http-client` is found but curl is not or too old
 
 4.4.0
 -----